メモリーアドレス表現部分の処理を分離
