+++ src/CMakeLists.txt
@@ -48,6 +48,7 @@ set(nask_parse_bnfc_SRCS
 set(nask_parse_front_SRCS
   front_end/front_config.cc
   front_end/front_parse.cc
+  front_end/front_memory_addr.cc
   front_end/front_inst.cc
   front_end/front_inst_jmp.cc
   front_end/front_inst_pseudo.cc)
